// src/app/main/common/models/login.model.ts
/**
 * @file: login.model.ts
 * @description: This file exports the interface used in login page
 * @author: Asish Das
 */
 export interface LoginSuccessFailModel {
    'success': string;
    'status code': number;
    'message': string;
    'name': string;
    'access': string;
    'refresh': string;

}